+++ src/App.js
@@ -1,5 +1,5 @@
 import './App.css';
-import Header from './components/header';
+import Header from './components/Header';
 import Form from './components/Form';
 import { useState } from 'react';
